% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AbundancePlot.R
\name{AbundancePlot-class}
\alias{AbundancePlot-class}
\title{Abundance plot}
\description{
Composite abundance profile of all features in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
object.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualisation:
\itemize{
\item \code{rank}, a string specifying the taxonomic rank to visualise
\item \code{add_legend}, a logical value indicating whether the legend should appear
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Constructor}{

\code{AbundancePlot(...)} creates an instance of a AbundancePlot class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\examples{
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

rowData(tse)$prevalence <- getPrevalence(tse,
                                         detection = 1/100,
                                         sort = FALSE,
                                         assay.type = "counts",
                                         as_relative = TRUE)
                                         
tse_genus <- mergeFeaturesByPrevalence(tse,
                                       rank = "Genus",
                                       prevalence = 50/100)
                                       
tse_genus <- addTaxonomyTree(tse_genus)
tse_genus <- transformAssay(tse_genus, method = "relabundance")

if (interactive()) {
  iSEE(tse_genus)
}

}
\author{
Giulio Benedetti
}
