% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/class-ColumnTreePlot.R
\docType{class}
\name{ColumnTreePlot-class}
\alias{ColumnTreePlot-class}
\alias{ColumnTreePlot}
\title{Column tree plot}
\value{
The \code{ColumnTreePlot(...)} constructor creates an instance of a ColumnTreePlot
class, where any slot and its value can be passed to \code{...} as a named
argument.
}
\description{
Hierarchical tree for the columns of a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
object. The tree represents the sample hierarchy of the study and gets stored
in the \code{\link[TreeSummarizedExperiment:rowLinks]{colTree}} slot of the
experiment object. The panel implements \code{\link[miaViz:plotTree]{plotColTree}}
to generate the plot.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualisation:
\itemize{
\item \code{layout}, a string specifying tree layout
\item \code{add_legend}, a logical indicating if color legend should appear.
\item \code{edge_colour_by}, a string specifying parameter to color lines by
when \code{colour_parameters = "Edge"}.
\item \code{edge_size_by}, a string specifying parameter to size lines by
when \code{size_parameters = "Edge"}.
\item \code{tip_colour_by}, a string specifying parameter to color tips by
when \code{colour_parameters = "Tip"}.
\item \code{tip_size_by}, a string specifying parameter to size tips by
when \code{size_parameters = "Tip"}.
\item \code{tip_shape_by}, a string specifying parameter to shape tips by
when \code{shape_parameters = "Tip"}.
\item \code{node_colour_by}, a string specifying parameter to color nodes by
when \code{colour_parameters = "Node"}.
\item \code{node_size_by}, a string specifying parameter to size nodes by
when \code{size_parameters = "Node"}.
\item \code{node_shape_by}, a string specifying parameter to shape nodes by
when \code{shape_parameters = "Node"}.
\item \code{order_tree}, a logical indicating if tree is ordered by
alphabetic order of taxonomic levels.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\examples{
# Import TreeSE
library(mia)
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Store panel into object
panel <- ColumnTreePlot()
# View some adjustable parameters
head(slotNames(panel))

# Launch iSEE with custom initial panel
if (interactive()) {
  iSEE(tse, initial = c(panel))
}

}
\author{
Giulio Benedetti
}
