% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-default.R
\docType{methods}
\name{iSEE-default}
\alias{iSEE-default}
\alias{iSEE}
\alias{iSEE,TreeSummarizedExperiment-method}
\title{iSEE layout for TreeSE}
\value{
The iSEE method for the TreeSE container returns a default set of panels
typically relevant for microbiome data. This configuration can be modified
by defining a different set of initial panels. By default, the visualised
panels include the following:
\itemize{
\item \code{RowDataTable()}
\item \code{ColumnDataTable()}
\item \code{RowTreePlot()}
\item \code{AbundancePlot()}
\item \code{AbundanceDensityPlot()}
\item \code{ReducedDimensionPlot()}
\item \code{ComplexHeatmapPlot()}
}
}
\description{
Panel configuration tuned to the specific properties of
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}.
}
\examples{
# Import TreeSE
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

# Agglomerate TreeSE by Genus
tse_genus <- agglomerateByRank(tse,
                               rank = "Genus",
                               onRankOnly = TRUE)

# Add relabundance assay
tse_genus <- transformAssay(tse_genus, method = "relabundance")

# Launch iSEE with custom initial panels
if (interactive()) {
  iSEE(tse_genus, initial = c(RowTreePlot(), AbundancePlot(), AbundanceDensityPlot()))
}

}
