% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEtree-package.R
\docType{package}
\name{iSEEtree-pkg}
\alias{iSEEtree}
\alias{iSEEtree-package}
\title{iSEE extension for the TreeSummarizedExperiment container}
\description{
\code{iSEEtree} is an extension of \pkg{iSEE} that provides panels for the
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
container, enabling the interactive visualisation of typical microbiome data.
}
\examples{
library("iSEEtree")
}
\author{
Giulio Benedetti \email{giulio.benedetti@utu.fi}
}
\keyword{internal}
