% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEEtree-package.R
\docType{package}
\name{iSEEtree}
\alias{iSEEtree-package}
\alias{iSEEtree}
\title{iSEE extension for the TreeSummarizedExperiment container}
\description{
\code{iSEEtree} is an extension of \pkg{iSEE} that provides panels for the
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
container, enabling the interactive visualisation of typical microbiome data.
The panel layout of iSEEtree is described in \code{\link{iSEE}}.
}
\examples{
library(iSEEtree)

}
\seealso{
\link[iSEE:iSEE]{iSEE} \link[mia:mia]{mia}
}
\author{
Giulio Benedetti \email{giulio.benedetti@utu.fi}
}
\keyword{internal}
