% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_RowTreePlot.R
\name{RowTreePlot-class}
\alias{RowTreePlot-class}
\title{Row tree plot}
\description{
A dimensionality reduction plot that dynamically recomputes the coordinates for the samples,
based on the selected subset of samples (and possibly features) in transmitting panels.
All samples in active and saved multiple selections are used here.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{layout}, a string specifying
\item \code{add_legend}, an integer scalar specifying
\item \code{colour_tip},
\item \code{colour_edge_by}, string indicating
\item \code{colour_tip},
\item \code{colour_tip_by},
}

In addition, this class inherits all slots from its parent \linkS4class{ColumnDotPlot},
\linkS4class{DotPlot} and \linkS4class{Panel} classes.
}

\section{Constructor}{

\code{RowTreePlot(...)} creates an instance of a RowTreePlot class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\examples{
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

rowData(tse)$prevalence <- getPrevalence(tse,
                                         detection = 1/100,
                                         sort = FALSE,
                                         assay.type = "counts",
                                         as_relative = TRUE)
                                         
tse_genus <- mergeFeaturesByPrevalence(tse,
                                       rank = "Genus",
                                       prevalence = 50/100)
                                       
tse_genus <- addTaxonomyTree(tse_genus)
tse_genus <- transformAssay(tse_genus, method = "relabundance")

if (interactive()) {
  iSEE(tse_genus)
}

}
\author{
Giulio Benedetti
}
