% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\alias{.check_all_panels}
\alias{.check_panel}
\title{iSEEtree utils}
\usage{
.check_all_panels(se, initial)

.check_panel(se, panel_list, panel_class, panel_fun, wtext)
}
\value{
\code{.check_panel} returns the input \code{initial} list of panels excluding
the checked panel if \code{panel_fun} is \code{NULL} or empty.
\code{.check_all_panels} applies \code{.check_panel} to multiple panels and
returns the a filtered version of \code{initial}.
}
\description{
Utility functions to check the existence of specific elements in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
that are compulsory when using certain panels.
}
\examples{
# Import libraries
library(mia)
library(TreeSummarizedExperiment)

# Import TreeSE
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Create list of panels
initial <- c(RowTreePlot(), ColumnTreePlot())
# If RowTreePlot is in initial, check whether rowLinks is defined
initial <- .check_panel(tse, initial, "RowTreePlot", rowLinks)
# If ColumnTreePlot is in initial, check whether colLinks is defined
initial <- .check_panel(tse, initial, "ColumnTreePlot", colLinks)

# View filtered list of panels
initial

}
\keyword{internal}
