% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AbundanceDensityPlot.R
\docType{methods}
\name{AbundanceDensityPlot}
\alias{AbundanceDensityPlot}
\alias{AbundanceDensityPlot-class}
\alias{initialize,AbundanceDensityPlot-method}
\title{Abundance density plot}
\description{
Density abundance profile of single features in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}.
The panel implements \code{\link[miaViz:plotAbundanceDensity]{plotAbundanceDensity}}
to generate the plot.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{layout}, a string specifying abundance layout (jitter, density or points).
\item \code{assay.type}, a string specifying the assay to visualize.
\item \code{n}, a number indicating the number of top taxa to visualize.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Constructor}{

\code{AbundanceDensityPlot(...)} creates an instance of an AbundanceDensityPlot
class, where any slot and its value can be passed to \code{...} as a named argument.
}

\examples{
# Import TreeSE
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

# Agglomerate TreeSE by Genus
tse_genus <- agglomerateByRank(tse,
                               rank = "Genus",
                               onRankOnly = TRUE)

# Add relabundance assay
tse_genus <- transformAssay(tse_genus, method = "relabundance")

# Launch iSEE
if (interactive()) {
  iSEE(tse_genus)
}

}
\author{
Giulio Benedetti
}
