% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-AbundancePlot.R
\docType{methods}
\name{AbundancePlot}
\alias{AbundancePlot}
\alias{AbundancePlot-class}
\alias{initialize,AbundancePlot-method}
\title{Abundance plot}
\description{
Composite abundance profile of all features in a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
object. The panel implements \code{\link[miaViz:plotAbundance]{plotAbundance}}
to generate the plot.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{rank}, a string specifying the taxonomic rank to visualize.
\item \code{add_legend}, a logical indicating if the color legend should appear.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Constructor}{

\code{AbundancePlot(...)} creates an instance of a AbundancePlot class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\examples{
# Import TreeSE
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

# Agglomerate TreeSE by Genus
tse_genus <- agglomerateByRank(tse,
                               rank = "Genus",
                               onRankOnly = TRUE)

# Launch iSEE
if (interactive()) {
  iSEE(tse_genus)
}

}
\author{
Giulio Benedetti
}
