% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iSEE-default.R
\name{iSEE-TreeSE}
\alias{iSEE-TreeSE}
\title{iSEE layout for TreeSE}
\description{
Panel configuration tuned to the specific properties of
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}.
}
\section{Default layout}{


The default configuration can be overwritten by defining a different set of
initial panels. By default, the visualized panels include the following:
\itemize{
\item \code{RowTreePlot()}
\item \code{AbundancePlot()}
\item \code{AbundanceDensityPlot()}
\item \code{ReducedDimensionPlot()}
\item \code{ComplexHeatmapPlot()}
}
}

\examples{
# Import TreeSE
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

# Agglomerate TreeSE by Genus
tse_genus <- mergeFeaturesByRank(tse,
                                 rank = "Genus",
                                 onRankOnly = TRUE)

# Add relabundance assay
tse_genus <- transformAssay(tse_genus, method = "relabundance")

# Launch iSEE with custom initial panels
if (interactive()) {
  iSEE(tse_genus, initial = c(RowTreePlot(), AbundancePlot(), AbundanceDensityPlot()))
}

}
