% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RowTreePlot.R
\name{RowTreePlot-class}
\alias{RowTreePlot-class}
\title{Row tree plot}
\description{
Hierarchical tree for the rows of a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
object. The tree can be produced with \code{\link[mia:taxonomy-methods]{addTaxonomyTree}}
and gets stored in the \code{\link[TreeSummarizedExperiment:rowLinks]{rowTree}}
slot of the experiment object. The panel implements \code{\link[miaViz:plotTree]{plotRowTree}}
to generate the plot.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{layout}, a string specifying tree layout
\item \code{add_legend}, a logical indicating if color legend should appear.
\item \code{colour_edge}, a string specifying color mapping for tree lines.
\item \code{colour_edge_by}, a string specifying parameter to color lines by
when \code{colour_edge = "Row data"}.
\item \code{colour_tip}, a string specifying color mapping for tree nodes.
\item \code{colour_tip_by}, a string specifying parameter to color nodes by
when \code{colour_tip = "Row data"}.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\section{Constructor}{

\code{RowTreePlot(...)} creates an instance of a RowTreePlot class,
where any slot and its value can be passed to \code{...} as a named argument.
}

\examples{
# Import TreeSE
library(mia)
data("GlobalPatterns", package = "mia")
tse <- GlobalPatterns

# Agglomerate TreeSE by Genus
tse_genus <- mergeFeaturesByRank(tse,
                                 rank = "Genus",
                                 onRankOnly = TRUE)

# Launch iSEE
if (interactive()) {
  iSEE(tse_genus)
}

}
\author{
Giulio Benedetti
}
