% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-RowTreePlot.R
\docType{methods}
\name{RowTreePlot}
\alias{RowTreePlot}
\alias{RowTreePlot-class}
\alias{initialize,RowTreePlot-method}
\title{Row tree plot}
\value{
The \code{RowTreePlot(...)} constructor creates an instance of a RowTreePlot
class, where any slot and its value can be passed to \code{...} as a named
argument.
}
\description{
Hierarchical tree for the rows of a
\code{\link[TreeSummarizedExperiment:TreeSummarizedExperiment-constructor]{TreeSummarizedExperiment}}
object. The tree can be produced with \code{\link[mia:taxonomy-methods]{addTaxonomyTree}}
and gets stored in the \code{\link[TreeSummarizedExperiment:rowLinks]{rowTree}}
slot of the experiment object. The panel implements \code{\link[miaViz:plotTree]{plotRowTree}}
to generate the plot.
}
\section{Slot overview}{

The following slots control the thresholds used in the visualization:
\itemize{
\item \code{layout}, a string specifying tree layout
\item \code{add_legend}, a logical indicating if color legend should appear.
\item \code{colour_edge}, a string specifying color mapping for tree lines.
\item \code{colour_edge_by}, a string specifying parameter to color lines by
when \code{colour_edge = "Row data"}.
\item \code{colour_tip}, a string specifying color mapping for tree nodes.
\item \code{colour_tip_by}, a string specifying parameter to color nodes by
when \code{colour_tip = "Row data"}.
\item \code{order_tree}, a logical indicating if tree is ordered by
alphabetic order of taxonomic levels.
}

In addition, this class inherits all slots from its parent \linkS4class{Panel} class.
}

\examples{
# Import TreeSE
library(mia)
data("Tengeler2020", package = "mia")
tse <- Tengeler2020

# Launch iSEE
if (interactive()) {
  iSEE(tse)
}

}
\author{
Giulio Benedetti
}
